# client/admin/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "PharmaCore",
    "description": "ERP Nhà thuốc đa quốc gia",
    "version": "v1"
  },
  "paths": {
...
    "/api/sales/customers": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerListItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/pos/lookup": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "barcode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priceType",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PosProductLookupDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosProductLookupDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosProductLookupDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/pos/stock": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "productUnitId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PosStockCheckDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosStockCheckDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosStockCheckDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/pos/stock/bulk": {
      "post": {
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosBulkStockRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PosBulkStockRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PosBulkStockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosStockCheckDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosStockCheckDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosStockCheckDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/pos/search": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priceType",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosProductSearchItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosProductSearchItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosProductSearchItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/pos/preview-allocation": {
      "post": {
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosAllocationPreviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PosAllocationPreviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PosAllocationPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PosAllocationPreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosAllocationPreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosAllocationPreviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/orders": {
      "get": {
        "tags": [
          "Sales"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesOrderListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesOrderListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesOrderListItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/orders/{id}": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftSaleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftSaleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftSaleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/orders/{id}/complete": {
      "post": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDraftSaleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDraftSaleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDraftSaleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/orders/{id}/cancel": {
      "post": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOrderDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/orders/{id}/returns": {
      "post": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleReturnRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleReturnRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleReturnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/returns": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReturnListItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/returns/{id}": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReturnDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/shift-summary": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SalesShiftSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesShiftSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesShiftSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerCode": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PosProductLookupDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "productUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "unitName": {
            "type": "string",
            "nullable": true
          },
          "conversionFactor": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "stockAvailable": {
            "type": "number",
            "format": "double"
          },
          "batchHints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PosBatchHintDto"
            },
            "nullable": true
          },
          "stockSourceLabel": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Nhập JWT token: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}